[feat/#2] 1203 잔여연차 계산, empNm, empNo 가져오는 setEmpInfo 구현

+++ src/main/java/com/supercoding/hrms/leave/domain/TblLeave.java
@@ -38,7 +38,7 @@ public class TblLeave {
     private String fileId; // 첨부파일 id
 
     @Column(name = "leave_duration")
-    private int leaveDuration; //휴가 기간
+    private Double leaveDuration; //휴가 기간
 
     @Column(name = "leave_status")
     private String leaveStatus; //상태
